refactor: move project user list item to separate file
